test: Deduplicate utilities for creating tables
