+++ crates/core/src/subscription/query.rs
@@ -217,38 +217,6 @@ mod tests {
     use spacetimedb_vm::dsl::{db_table, mem_table, scalar};
     use spacetimedb_vm::operator::OpCmp;
 
-    fn create_table(
-        db: &RelationalDB,
-        tx: &mut MutTx,
-        name: &str,
-        schema: &[(&str, AlgebraicType)],
-        indexes: &[(ColId, &str)],
-    ) -> ResultTest<TableId> {
-        let table_name = name.to_string();
-        let table_type = StTableType::User;
-        let table_access = StAccess::Public;
-
-        let columns = schema
-            .iter()
-            .map(|(col_name, col_type)| ColumnDef {
-                col_name: col_name.to_string(),
-                col_type: col_type.clone(),
-            })
-            .collect_vec();
-
-        let indexes = indexes
-            .iter()
-            .map(|(col_id, index_name)| IndexDef::btree(index_name.to_string(), *col_id, false))
-            .collect_vec();
-
-        let schema = TableDef::new(table_name, columns)
-            .with_indexes(indexes)
-            .with_type(table_type)
-            .with_access(table_access);
-
-        Ok(db.create_table(tx, schema)?)
-    }
-
     fn insert_op(table_id: TableId, table_name: &str, row: ProductValue) -> DatabaseTableUpdate {
         let row_pk = row.to_data_key().to_bytes();
         DatabaseTableUpdate {
@@ -474,13 +442,13 @@ mod tests {
     #[test]
     fn test_eval_incr_for_index_scan() -> ResultTest<()> {
         let (db, _tmp) = make_test_db()?;
-        let mut tx = db.begin_mut_tx();
 
         // Create table [test] with index on [b]
         let schema = &[("a", AlgebraicType::U64), ("b", AlgebraicType::U64)];
         let indexes = &[(1.into(), "b")];
-        let table_id = create_table(&db, &mut tx, "test", schema, indexes)?;
+        let table_id = db.create_table_for_test("test", schema, indexes)?;
 
+        let mut tx = db.begin_mut_tx();
         let mut ops = Vec::new();
         for i in 0u64..9u64 {
             let row = product!(i, i);
@@ -542,12 +510,10 @@ mod tests {
     }
 
     fn run_eval_incr_for_index_join(db: RelationalDB) -> ResultTest<()> {
-        let mut tx = db.begin_mut_tx();
-
         // Create table [lhs] with index on [id]
         let schema = &[("id", AlgebraicType::I32), ("x", AlgebraicType::I32)];
         let indexes = &[(0.into(), "id")];
-        let lhs_id = create_table(&db, &mut tx, "lhs", schema, indexes)?;
+        let lhs_id = db.create_table_for_test("lhs", schema, indexes)?;
 
         // Create table [rhs] with index on [id]
         let schema = &[
@@ -556,7 +522,9 @@ mod tests {
             ("y", AlgebraicType::I32),
         ];
         let indexes = &[(1.into(), "id")];
-        let rhs_id = create_table(&db, &mut tx, "rhs", schema, indexes)?;
+        let rhs_id = db.create_table_for_test("rhs", schema, indexes)?;
+
+        let mut tx = db.begin_mut_tx();
 
         // Insert into lhs
         for i in 0..5 {
@@ -967,7 +935,6 @@ mod tests {
     #[test]
     fn test_subscribe_sql() -> ResultTest<()> {
         let (db, _tmp_dir) = make_test_db()?;
-        let mut tx = db.begin_mut_tx();
 
         // Create table [MobileEntityState]
         let schema = &[
@@ -985,7 +952,7 @@ mod tests {
             (1.into(), "location_x"),
             (2.into(), "location_z"),
         ];
-        create_table(&db, &mut tx, "MobileEntityState", schema, indexes)?;
+        db.create_table_for_test("MobileEntityState", schema, indexes)?;
 
         // Create table [EnemyState]
         let schema = &[
@@ -996,8 +963,7 @@ mod tests {
             ("direction", AlgebraicType::I32),
         ];
         let indexes = &[(0.into(), "entity_id")];
-        create_table(&db, &mut tx, "EnemyState", schema, indexes)?;
-        db.commit_tx(&ExecutionContext::default(), tx)?;
+        db.create_table_for_test("EnemyState", schema, indexes)?;
 
         let sql_insert = "\
         insert into MobileEntityState (entity_id, location_x, location_z, destination_x, destination_z, is_running, timestamp, dimension) values (1, 96001, 96001, 96001, 1867045146, false, 17167179743690094247, 3926297397);\
@@ -1084,22 +1050,20 @@ mod tests {
     #[test]
     fn test_classify() -> ResultTest<()> {
         let (db, _tmp_dir) = make_test_db()?;
-        let mut tx = db.begin_mut_tx();
 
         // Create table [plain]
         let schema = &[("id", AlgebraicType::U64)];
-        create_table(&db, &mut tx, "plain", schema, &[])?;
+        db.create_table_for_test("plain", schema, &[])?;
 
         // Create table [lhs] with indexes on [id] and [x]
         let schema = &[("id", AlgebraicType::U64), ("x", AlgebraicType::I32)];
         let indexes = &[(ColId(0), "id"), (ColId(1), "x")];
-        create_table(&db, &mut tx, "lhs", schema, indexes)?;
+        db.create_table_for_test("lhs", schema, indexes)?;
 
         // Create table [rhs] with indexes on [id] and [y]
         let schema = &[("id", AlgebraicType::U64), ("y", AlgebraicType::I32)];
         let indexes = &[(ColId(0), "id"), (ColId(1), "y")];
-        create_table(&db, &mut tx, "rhs", schema, indexes)?;
-        db.commit_tx(&ExecutionContext::default(), tx)?;
+        db.create_table_for_test("rhs", schema, indexes)?;
 
         let tx = db.begin_tx();
         let auth = AuthCtx::for_testing();
